添加功能：
1、Server中计算上传当前文件耗时。

2、将Client上传文件功能具体实现转移到FileToSend对象中，实现了通过子线程进行数据上传，不会阻塞LogicSystem中消息队列的
信息处理；也减轻LogicSystem的处理压力。

3、Client端在重传缺包时，不再与上传文件时使用同一个文件流，以防止读取缺包文件数据时，与之前的文件流光标位置产生冲突。（Server端只会
按照顺序进行持久化，所以不会产生这个问题）

4、Server端在同一时间内只会发送一次重传，如果在一个重传周期内再次请求重传，后面那一次重传将不会被允许。每接受到seq%100==0的
seq包时，会进行一次缺包检测。

5、将几乎所有unsigned int 改为uintmax_t

6、添加显示Server当前所利用存储盘的剩余大小。后续可以根据这一大小进行能否上传的判断。

+++ client.h
@@ -74,6 +74,6 @@ private:
     // 用于断电功能的状态值
     bool is_Uploading; //Client正在上传文件到Server 用于析构函数中判断状态，是否进行断点记载
     bool is_Downloading; //Client正在下载文件到Server
-    unsigned int m_CurrentSeq; //这里要记录每一个文件的最近文件号
+    uintmax_t m_CurrentSeq; //这里要记录每一个文件的最近文件号
     // 当前传输的文件路径在m_FilesToSend里找
 };
